brsr_generator/requirements.txt
requests
python-dotenv
python-docx
PyPDF2
openpyxl 
crewai 